:recycle: Refactor CoordinatorController (#127)
Updated booking now always has status CHANDED.
If Coordinator tries to grant Interviewer to Interviewer exception is thrown.
Refactor dashboard endpoint to display days in an appropriate format.
Add a method for converting weekNum + dayOfWeek to LocalDate and write a test for it. (Added it in order to implement the dashboard).